correcciones de migracion
correccion en proveedor y datos

+++ data/creaciontablasyesquema.sql
@@ -1,9 +1,10 @@
 --use GD2C2019
-
+--GO
 --CREACION DE ESQUEMA
 
---IF OBJECT_ID('LIL_MIX.usuario') IS NOT NULL
-  --DROP TABLE LIL_MIX.usuario
+IF OBJECT_ID('LIL_MIX.usuario') IS NOT NULL
+  DROP TABLE LIL_MIX.usuario
+GO
 
 CREATE SCHEMA LIL_MIX AUTHORIZATION  gdCupon2019 
 
@@ -28,59 +29,75 @@ GO
 
 IF OBJECT_ID('LIL_MIX.usuario') IS NOT NULL
   DROP TABLE LIL_MIX.usuario
+  GO
 
 IF OBJECT_ID('LIL_MIX.direccion') IS NOT NULL
   DROP TABLE LIL_MIX.direccion
+  GO
 
 IF OBJECT_ID('LIL_MIX.proveedor') IS NOT NULL
   DROP TABLE LIL_MIX.proveedor
+  GO
 
 IF OBJECT_ID('LIL_MIX.oferta') IS NOT NULL
   DROP TABLE LIL_MIX.oferta
+ GO
 
 IF OBJECT_ID('LIL_MIX.cliente') IS NOT NULL
   DROP TABLE LIL_MIX.cliente
+ GO
 
 IF OBJECT_ID('LIL_MIX.compra') IS NOT NULL
   DROP TABLE LIL_MIX.compra
+ GO
 
 IF OBJECT_ID('LIL_MIX.factura') IS NOT NULL
   DROP TABLE LIL_MIX.factura
+ GO
 
 IF OBJECT_ID('LIL_MIX.cupon') IS NOT NULL
   DROP TABLE LIL_MIX.cupon
+ GO
 
 IF OBJECT_ID('LIL_MIX.rol') IS NOT NULL
   DROP TABLE LIL_MIX.rol
+ GO
 
 IF OBJECT_ID('LIL_MIX.funcionalidad') IS NOT NULL
   DROP TABLE LIL_MIX.funcionalidad
+ GO
+
 
 IF OBJECT_ID('LIL_MIX.rolxusuario') IS NOT NULL
   DROP TABLE LIL_MIX.rolxusuario
+ GO
 
 IF OBJECT_ID('LIL_MIX.funcionalidadxrol') IS NOT NULL
   DROP TABLE LIL_MIX.funcionalidadxrol
+ GO
 
 IF OBJECT_ID('LIL_MIX.tarjeta') IS NOT NULL
   DROP TABLE LIL_MIX.tarjeta
+ GO
 
 IF OBJECT_ID('LIL_MIX.tipoDePago') IS NOT NULL
   DROP TABLE LIL_MIX.tipoDePago
+ GO
 
 IF OBJECT_ID('LIL_MIX.cargaDeCredito') IS NOT NULL
   DROP TABLE LIL_MIX.cargaDeCredito
-  
+ GO 
+
 --CREACION DE TABLAS 
 
-CREATE TABLE LIL_MIX.usuario ( usuario_id INT NOT NULL IDENTITY(1000,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.usuario (usuario_id INT NOT NULL IDENTITY(1000,1) PRIMARY KEY,
 			       usuario_nombre VARCHAR(255) NOT NULL UNIQUE,
 	                       usuario_password VARCHAR(255) NOT NULL,
 			       usuario_intentos TINYINT DEFAULT 0,
 			       usuario_habilitado BIT DEFAULT 1
 			      )
 									   					 
-CREATE TABLE LIL_MIX.direccion ( direccion_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.direccion (direccion_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				 direccion_calle VARCHAR(255),
 				 direccion_piso TINYINT,
 				 direccion_dpto CHAR(1),
@@ -96,11 +113,11 @@ CREATE TABLE LIL_MIX.proveedor( proveedor_id INT NOT NULL IDENTITY (1,1) PRIMARY
 				proveedor_cp SMALLINT,
 				proveedor_nombre_contacto VARCHAR(255),
 				proveedor_rs VARCHAR(255) UNIQUE,
-				proveedor_habilitado BIT,
+				proveedor_habilitado BIT DEFAULT 1,
 				proveedor_usuario_id INT FOREIGN KEY REFERENCES LIL_MIX.usuario(usuario_id)
 				)
 
-CREATE TABLE LIL_MIX.oferta ( 	oferta_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.oferta (oferta_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				oferta_codigo VARCHAR(255) NOT NULL,
 				oferta_precio_oferta INT NOT NULL,
 				oferta_precio_lista INT NOT NULL,
@@ -112,7 +129,7 @@ CREATE TABLE LIL_MIX.oferta ( 	oferta_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				oferta_restriccion_compra TINYINT NOT NULL
 				)
 
-CREATE TABLE LIL_MIX.cliente ( 	cliente_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY, --el cliente id comenzará en 1 y se incrementará en 1 a medida que se vayan agregando nuevos clientes
+CREATE TABLE LIL_MIX.cliente (cliente_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY, --el cliente id comenzará en 1 y se incrementará en 1 a medida que se vayan agregando nuevos clientes
 				cliente_nombre VARCHAR(255) ,
 				cliente_apellido VARCHAR(255) , 
 				cliente_direccion_id INT FOREIGN KEY REFERENCES LIL_MIX.direccion(direccion_id),
@@ -122,11 +139,11 @@ CREATE TABLE LIL_MIX.cliente ( 	cliente_id INT NOT NULL IDENTITY(1,1) PRIMARY KE
 				cliente_cp SMALLINT ,
 				cliente_dni INT,
 				cliente_credito BIGINT,
-				cliente_habilitado BIT,
+				cliente_habilitado BIT DEFAULT 1,
 				cliente_usuario_id INT FOREIGN KEY REFERENCES LIL_MIX.usuario(usuario_id)
 				) 
 
-CREATE TABLE LIL_MIX.compra ( 	compra_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.compra (compra_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				compra_oferta_numero INT FOREIGN KEY REFERENCES LIL_MIX.oferta(oferta_id),
 				compra_oferta_descr VARCHAR(255),
 				compra_cliente_id INT FOREIGN KEY REFERENCES LIL_MIX.cliente(cliente_id),
@@ -134,23 +151,23 @@ CREATE TABLE LIL_MIX.compra ( 	compra_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				compra_fecha DATETIME
 				)
 
-CREATE TABLE LIL_MIX.factura ( 	factura_id INT NOT NULL IDENTITY(1000,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.factura (factura_id INT NOT NULL IDENTITY(1000,1) PRIMARY KEY,
 				factura_proveedor_id INT FOREIGN KEY REFERENCES LIL_MIX.proveedor(proveedor_id),
 				factura_fecha_inicio DATETIME,
 				factura_fecha_fin DATETIME,
 				factura_importe FLOAT
 				)
 
-CREATE TABLE LIL_MIX.cupon ( 	cupon_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.cupon (cupon_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				cupon_fecha_vencimiento DATETIME,
 				cupon_fecha_consumo DATETIME,
 				cupon_compra_id INT FOREIGN KEY REFERENCES LIL_MIX.compra(compra_id),
 				cupon_cliente_id INT FOREIGN KEY REFERENCES LIL_MIX.cliente(cliente_id)
 				)
 
-CREATE TABLE LIL_MIX.rol ( rol_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.rol (rol_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 			   rol_nombre VARCHAR(30) NOT NULL,
-			   rol_habilitado BIT
+			   rol_habilitado BIT DEFAULT 1
 			   )
 
 CREATE TABLE LIL_MIX.funcionalidad ( funcionalidad_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
@@ -164,24 +181,24 @@ CREATE TABLE LIL_MIX.rolxusuario ( rol_id INT,
 				   FOREIGN KEY (usuario_id) REFERENCES LIL_MIX.usuario(usuario_id)
 				   )
 
-CREATE TABLE LIL_MIX.funcionalidadxrol ( rol_id INT,
+CREATE TABLE LIL_MIX.funcionalidadxrol (rol_id INT,
 					 funcionalidad_id INT,
 					 PRIMARY KEY (rol_id, funcionalidad_id),
 					 FOREIGN KEY (rol_id) REFERENCES LIL_MIX.rol(rol_id),
 					 FOREIGN KEY (funcionalidad_id) REFERENCES LIL_MIX.funcionalidad(funcionalidad_id)
 					 )
 
-CREATE TABLE LIL_MIX.tarjeta ( 	tarjeta_numero BIGINT NOT NULL PRIMARY KEY,
+CREATE TABLE LIL_MIX.tarjeta (tarjeta_numero BIGINT NOT NULL PRIMARY KEY,
 				tarjeta_tipo VARCHAR(30),
 				tarjeta_fecha_vencimiento DATETIME,
 				tarjeta_id_cliente INT
 				)
 
-CREATE TABLE LIL_MIX.tipoDePago ( tipo_de_pago_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY, --1, 2, 3
+CREATE TABLE LIL_MIX.tipoDePago (tipo_de_pago_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY, --1, 2, 3
 				  tipo_de_pago_descripcion VARCHAR(30) --EFECTIVO, CREDITO O DEBITO
 				 )
 
-CREATE TABLE LIL_MIX.cargaDeCredito ( carga_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.cargaDeCredito (carga_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				      carga_fecha DATETIME NOT NULL,
 				      carga_id_cliente INT NOT NULL FOREIGN KEY REFERENCES LIL_MIX.cliente(cliente_id),
 				      carga_tipo_de_pago INT NOT NULL FOREIGN KEY REFERENCES LIL_MIX.tipoDePago(tipo_de_pago_id), --TODAVIA NO EJECUTAMOS
@@ -189,7 +206,7 @@ CREATE TABLE LIL_MIX.cargaDeCredito ( carga_id INT NOT NULL IDENTITY(1,1) PRIMAR
 				      carga_tarjeta_numero BIGINT FOREIGN KEY REFERENCES LIL_MIX.tarjeta(tarjeta_numero) 
 				     )
 
-CREATE TABLE LIL_MIX.semestre ( semestre_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
+CREATE TABLE LIL_MIX.semestre (semestre_id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
 				semestre_fecha_inicio VARCHAR(5),
 				semestre_fecha_fin VARCHAR(5)
 				)
